pluralscan-webapp/pyproject.toml
[tool.poetry]
name = "pluralscan-web"
version = "0.1.0"
description = ""
authors = ["Luidgi Gromat <gromat.luidgi@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9"
Django = "^4.0.5"
whitenoise = "^6.1.0"
djangorestframework = "^3.13.1"
Markdown = "^3.3.7"
pymongo = "^4.1.1"
gunicorn = "^20.1.0"
requests = "^2.28.0"
django-cors-headers = "^3.13.0"
django-environ = "^0.9.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pytest-drf = "^1.1.2"
setuptools = "^63.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
